Rename client-redirects file
